Fix deadlock in win implementation of io_service
Due to the fact that socket map was cleared in disconnect method of tcp_client and completion port was signaled only if socket map was not empty all io threads were waiting forever to get the signal. This would potentially lead to a deadlock on shutdown in case redis client would run in a separate thread that would properly wait for destructor to finis.